cc: one more long->int32 in pswt.c for tests/c/misc/switch_hex.c
This was found by bisecting pool.8 difference when compiling
lib_core/libc/port/pool.c with 8c -m32 or -m64

+++ compilers/cc/pswt.c
@@ -53,7 +53,9 @@ doswit(Node *n)
 		if(isv)
 			q->val = c->val;
 		else
-			q->val = (long)c->val;	/* cast ensures correct value for 32-bit switch on 64-bit architecture */
+            //old: was (long) but actually not 32 bits on 64bits arch!
+            // which led to different switch order in tests/c/misc/switch_hex.c
+			q->val = (int32)c->val;	/* cast ensures correct value for 32-bit switch on 64-bit architecture */
 		q++;
 	}
 	qsort(iq, nc, sizeof(C1), swcmp);
@@ -105,7 +107,7 @@ doswit(Node *n)
 		for(j = i; j < nc; j++){
 			if((iq[j].val>>32) != iqh[nh].val)
 				break;
-			q->val = (long)iq[j].val;
+			q->val = (int32)iq[j].val;
 			q->label = iq[j].label;
 			q++;
 		}
